Uploading initial project files
Initial check-in of the project. It contains the interface to be implemented for the keyboard device + documenting comments taken straight from the arduino docs, commented out arduino sketch to run an arduino as a keyboard with a few example key presses and required Visual Studio project files if you're using the Arduino plug-in.