Adding windows as a platform + placeholder keyboard function definitions
This adds a second Visual Studio Solution which includes the functional
files of the arduino keyboard. Also added wrapper macros for a bunch of
the arduino functions so the keyboard can be run on windows if needed.
The intention is, development/logic tests etc, can be carried out on
windows, arduino platform, or arduino platform as a USB device using
#defines to make these decisions and forward calls to the right place.